@@ -1,18 +1,15 @@
 {
   "name": "human-signals",
-  "version": "2.1.0",
-  "main": "build/src/main.js",
+  "version": "3.0.1",
+  "type": "module",
+  "exports": "./build/src/main.js",
+  "main": "./build/src/main.js",
   "files": [
-    "build/src",
-    "!~"
+    "build/src/**/*.{js,ts,map,json,sh,md}",
+    "examples/**/*.{js,ts,map,json,sh,md}"
   ],
   "scripts": {
     "test": "gulp test"
-  },
-  "husky": {
-    "hooks": {
-      "pre-push": "gulp check --full"
-    }
   },
   "description": "Human-friendly process signals",
   "keywords": [
@@ -49,16 +46,12 @@
     "test": "test"
   },
   "types": "build/src/main.d.ts",
-  "dependencies": {},
   "devDependencies": {
-    "@ehmicky/dev-tasks": "^0.31.9",
-    "ajv": "^6.12.0",
-    "ava": "^3.5.0",
-    "gulp": "^4.0.2",
-    "husky": "^4.2.3",
-    "test-each": "^2.0.0"
+    "@ehmicky/dev-tasks": "^1.0.34",
+    "ajv": "^6.12.6",
+    "test-each": "^3.0.1"
   },
   "engines": {
-    "node": ">=10.17.0"
+    "node": ">=12.20.0"
   }
 }
