@@ -1,6 +1,6 @@
 {
 	"name": "execa",
-	"version": "4.1.0",
+	"version": "6.1.0",
 	"description": "Process execution for humans",
 	"license": "MIT",
 	"repository": "sindresorhus/execa",
@@ -10,11 +10,13 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd"
+		"test": "xo && c8 ava && tsd"
 	},
 	"files": [
 		"index.js",
@@ -39,33 +41,43 @@
 		"local"
 	],
 	"dependencies": {
-		"cross-spawn": "^7.0.0",
-		"get-stream": "^5.0.0",
-		"human-signals": "^1.1.1",
-		"is-stream": "^2.0.0",
+		"cross-spawn": "^7.0.3",
+		"get-stream": "^6.0.1",
+		"human-signals": "^3.0.1",
+		"is-stream": "^3.0.0",
 		"merge-stream": "^2.0.0",
-		"npm-run-path": "^4.0.0",
-		"onetime": "^5.1.0",
-		"signal-exit": "^3.0.2",
-		"strip-final-newline": "^2.0.0"
+		"npm-run-path": "^5.1.0",
+		"onetime": "^6.0.0",
+		"signal-exit": "^3.0.7",
+		"strip-final-newline": "^3.0.0"
 	},
 	"devDependencies": {
-		"@types/node": "^12.12.18",
-		"ava": "^2.1.0",
-		"coveralls": "^3.0.9",
-		"get-node": "^6.6.0",
+		"@types/node": "^17.0.17",
+		"ava": "^4.0.1",
+		"c8": "^7.11.0",
+		"get-node": "^12.0.0",
 		"is-running": "^2.1.0",
-		"nyc": "^14.1.1",
-		"p-event": "^4.1.0",
-		"tempfile": "^3.0.0",
-		"tsd": "^0.11.0",
-		"xo": "^0.25.3"
+		"p-event": "^5.0.1",
+		"semver": "^7.3.5",
+		"tempfile": "^4.0.0",
+		"tsd": "^0.19.1",
+		"xo": "^0.48.0"
 	},
-	"nyc": {
+	"c8": {
+		"reporter": [
+			"text",
+			"lcov"
+		],
 		"exclude": [
 			"**/fixtures/**",
 			"**/test.js",
 			"**/test/**"
 		]
+	},
+	"xo": {
+		"rules": {
+			"unicorn/no-empty-file": "off",
+			"@typescript-eslint/ban-types": "off"
+		}
 	}
 }
