@@ -1,16 +1,16 @@
 import { constants } from 'os'
 
-import { getSignals } from './signals.js'
 import { SIGRTMAX } from './realtime.js'
+import { getSignals } from './signals.js'
 
 // Retrieve `signalsByName`, an object mapping signal name to signal properties.
 // We make sure the object is sorted by `number`.
-const getSignalsByName = function() {
+const getSignalsByName = function () {
   const signals = getSignals()
   return signals.reduce(getSignalByName, {})
 }
 
-const getSignalByName = function(
+const getSignalByName = function (
   signalByNameMemo,
   { name, number, description, supported, action, forced, standard },
 ) {
@@ -25,7 +25,7 @@ export const signalsByName = getSignalsByName()
 // Retrieve `signalsByNumber`, an object mapping signal number to signal
 // properties.
 // We make sure the object is sorted by `number`.
-const getSignalsByNumber = function() {
+const getSignalsByNumber = function () {
   const signals = getSignals()
   const length = SIGRTMAX + 1
   const signalsA = Array.from({ length }, (value, number) =>
@@ -34,7 +34,7 @@ const getSignalsByNumber = function() {
   return Object.assign({}, ...signalsA)
 }
 
-const getSignalByNumber = function(number, signals) {
+const getSignalByNumber = function (number, signals) {
   const signal = findSignalByNumber(number, signals)
 
   if (signal === undefined) {
@@ -57,14 +57,14 @@ const getSignalByNumber = function(number, signals) {
 
 // Several signals might end up sharing the same number because of OS-specific
 // numbers, in which case those prevail.
-const findSignalByNumber = function(number, signals) {
+const findSignalByNumber = function (number, signals) {
   const signal = signals.find(({ name }) => constants.signals[name] === number)
 
   if (signal !== undefined) {
     return signal
   }
 
-  return signals.find(signalA => signalA.number === number)
+  return signals.find((signalA) => signalA.number === number)
 }
 
 export const signalsByNumber = getSignalsByNumber()
