@@ -1,21 +1,31 @@
+import {Buffer} from 'node:buffer';
+// For some reason a default import of `process` causes
+// `process.stdin`, `process.stderr`, and `process.stdout`
+// to get treated as `any` by `@typescript-eslint/no-unsafe-assignment`.
+import * as process from 'node:process';
+import {Readable as ReadableStream} from 'node:stream';
 import {expectType, expectError} from 'tsd';
-import {Readable as ReadableStream} from 'stream'
-import execa = require('.');
 import {
+	execa,
+	execaSync,
+	execaCommand,
+	execaCommandSync,
+	execaNode,
 	ExecaReturnValue,
 	ExecaChildProcess,
 	ExecaError,
 	ExecaSyncReturnValue,
-	ExecaSyncError
-} from '.';
+	ExecaSyncError,
+} from './index.js';
 
 try {
 	const execaPromise = execa('unicorns');
 	execaPromise.cancel();
-	expectType<ReadableStream | undefined>(execaPromise.all)
+	expectType<ReadableStream | undefined>(execaPromise.all);
 
 	const unicornsResult = await execaPromise;
 	expectType<string>(unicornsResult.command);
+	expectType<string>(unicornsResult.escapedCommand);
 	expectType<number>(unicornsResult.exitCode);
 	expectType<string>(unicornsResult.stdout);
 	expectType<string>(unicornsResult.stderr);
@@ -26,8 +36,8 @@ try {
 	expectType<boolean>(unicornsResult.killed);
 	expectType<string | undefined>(unicornsResult.signal);
 	expectType<string | undefined>(unicornsResult.signalDescription);
-} catch (error) {
-	const execaError: ExecaError = error;
+} catch (error: unknown) {
+	const execaError = error as ExecaError;
 
 	expectType<string>(execaError.message);
 	expectType<number>(execaError.exitCode);
@@ -45,8 +55,9 @@ try {
 }
 
 try {
-	const unicornsResult = execa.sync('unicorns');
+	const unicornsResult = execaSync('unicorns');
 	expectType<string>(unicornsResult.command);
+	expectType<string>(unicornsResult.escapedCommand);
 	expectType<number>(unicornsResult.exitCode);
 	expectType<string>(unicornsResult.stdout);
 	expectType<string>(unicornsResult.stderr);
@@ -57,8 +68,8 @@ try {
 	expectType<boolean>(unicornsResult.killed);
 	expectType<string | undefined>(unicornsResult.signal);
 	expectType<string | undefined>(unicornsResult.signalDescription);
-} catch (error) {
-	const execaError: ExecaSyncError = error;
+} catch (error: unknown) {
+	const execaError = error as ExecaSyncError;
 
 	expectType<string>(execaError.message);
 	expectType<number>(execaError.exitCode);
@@ -75,9 +86,11 @@ try {
 	expectType<string | undefined>(execaError.originalMessage);
 }
 
+/* eslint-disable @typescript-eslint/no-floating-promises */
 execa('unicorns', {cleanup: false});
 execa('unicorns', {preferLocal: false});
 execa('unicorns', {localDir: '.'});
+execa('unicorns', {localDir: new URL('file:///test')});
 execa('unicorns', {execPath: '/path'});
 execa('unicorns', {buffer: false});
 execa('unicorns', {input: ''});
@@ -109,13 +122,15 @@ execa('unicorns', {reject: false});
 execa('unicorns', {stripFinalNewline: false});
 execa('unicorns', {extendEnv: false});
 execa('unicorns', {cwd: '.'});
+execa('unicorns', {cwd: new URL('file:///test')});
+// eslint-disable-next-line @typescript-eslint/naming-convention
 execa('unicorns', {env: {PATH: ''}});
 execa('unicorns', {argv0: ''});
 execa('unicorns', {stdio: 'pipe'});
 execa('unicorns', {stdio: 'ignore'});
 execa('unicorns', {stdio: 'inherit'});
 execa('unicorns', {
-	stdio: ['pipe', 'ipc', 'ignore', 'inherit', process.stdin, 1, undefined]
+	stdio: ['pipe', 'ipc', 'ignore', 'inherit', process.stdin, 1, undefined],
 });
 execa('unicorns', {serialization: 'advanced'});
 execa('unicorns', {detached: true});
@@ -127,8 +142,10 @@ execa('unicorns', {timeout: 1000});
 execa('unicorns', {maxBuffer: 1000});
 execa('unicorns', {killSignal: 'SIGTERM'});
 execa('unicorns', {killSignal: 9});
+execa('unicorns', {signal: new AbortController().signal});
 execa('unicorns', {windowsVerbatimArguments: true});
 execa('unicorns', {windowsHide: false});
+/* eslint-enable @typescript-eslint/no-floating-promises */
 execa('unicorns').kill();
 execa('unicorns').kill('SIGKILL');
 execa('unicorns').kill(undefined);
@@ -137,55 +154,55 @@ execa('unicorns').kill('SIGKILL', {forceKillAfterTimeout: false});
 execa('unicorns').kill('SIGKILL', {forceKillAfterTimeout: 42});
 execa('unicorns').kill('SIGKILL', {forceKillAfterTimeout: undefined});
 
-expectType<ExecaChildProcess<string>>(execa('unicorns'));
-expectType<ExecaReturnValue<string>>(await execa('unicorns'));
-expectType<ExecaReturnValue<string>>(
-	await execa('unicorns', {encoding: 'utf8'})
+expectType<ExecaChildProcess>(execa('unicorns'));
+expectType<ExecaReturnValue>(await execa('unicorns'));
+expectType<ExecaReturnValue>(
+	await execa('unicorns', {encoding: 'utf8'}),
 );
 expectType<ExecaReturnValue<Buffer>>(await execa('unicorns', {encoding: null}));
-expectType<ExecaReturnValue<string>>(
-	await execa('unicorns', ['foo'], {encoding: 'utf8'})
+expectType<ExecaReturnValue>(
+	await execa('unicorns', ['foo'], {encoding: 'utf8'}),
 );
 expectType<ExecaReturnValue<Buffer>>(
-	await execa('unicorns', ['foo'], {encoding: null})
+	await execa('unicorns', ['foo'], {encoding: null}),
 );
 
-expectType<ExecaSyncReturnValue<string>>(execa.sync('unicorns'));
-expectType<ExecaSyncReturnValue<string>>(
-	execa.sync('unicorns', {encoding: 'utf8'})
+expectType<ExecaSyncReturnValue>(execaSync('unicorns'));
+expectType<ExecaSyncReturnValue>(
+	execaSync('unicorns', {encoding: 'utf8'}),
 );
 expectType<ExecaSyncReturnValue<Buffer>>(
-	execa.sync('unicorns', {encoding: null})
+	execaSync('unicorns', {encoding: null}),
 );
-expectType<ExecaSyncReturnValue<string>>(
-	execa.sync('unicorns', ['foo'], {encoding: 'utf8'})
+expectType<ExecaSyncReturnValue>(
+	execaSync('unicorns', ['foo'], {encoding: 'utf8'}),
 );
 expectType<ExecaSyncReturnValue<Buffer>>(
-	execa.sync('unicorns', ['foo'], {encoding: null})
+	execaSync('unicorns', ['foo'], {encoding: null}),
 );
 
-expectType<ExecaChildProcess<string>>(execa.command('unicorns'));
-expectType<ExecaReturnValue<string>>(await execa.command('unicorns'));
-expectType<ExecaReturnValue<string>>(await execa.command('unicorns', {encoding: 'utf8'}));
-expectType<ExecaReturnValue<Buffer>>(await execa.command('unicorns', {encoding: null}));
-expectType<ExecaReturnValue<string>>(await execa.command('unicorns foo', {encoding: 'utf8'}));
-expectType<ExecaReturnValue<Buffer>>(await execa.command('unicorns foo', {encoding: null}));
+expectType<ExecaChildProcess>(execaCommand('unicorns'));
+expectType<ExecaReturnValue>(await execaCommand('unicorns'));
+expectType<ExecaReturnValue>(await execaCommand('unicorns', {encoding: 'utf8'}));
+expectType<ExecaReturnValue<Buffer>>(await execaCommand('unicorns', {encoding: null}));
+expectType<ExecaReturnValue>(await execaCommand('unicorns foo', {encoding: 'utf8'}));
+expectType<ExecaReturnValue<Buffer>>(await execaCommand('unicorns foo', {encoding: null}));
 
-expectType<ExecaSyncReturnValue<string>>(execa.commandSync('unicorns'));
-expectType<ExecaSyncReturnValue<string>>(execa.commandSync('unicorns', {encoding: 'utf8'}));
-expectType<ExecaSyncReturnValue<Buffer>>(execa.commandSync('unicorns', {encoding: null}));
-expectType<ExecaSyncReturnValue<string>>(execa.commandSync('unicorns foo', {encoding: 'utf8'}));
-expectType<ExecaSyncReturnValue<Buffer>>(execa.commandSync('unicorns foo', {encoding: null}));
+expectType<ExecaSyncReturnValue>(execaCommandSync('unicorns'));
+expectType<ExecaSyncReturnValue>(execaCommandSync('unicorns', {encoding: 'utf8'}));
+expectType<ExecaSyncReturnValue<Buffer>>(execaCommandSync('unicorns', {encoding: null}));
+expectType<ExecaSyncReturnValue>(execaCommandSync('unicorns foo', {encoding: 'utf8'}));
+expectType<ExecaSyncReturnValue<Buffer>>(execaCommandSync('unicorns foo', {encoding: null}));
 
-expectType<ExecaChildProcess<string>>(execa.node('unicorns'));
-expectType<ExecaReturnValue<string>>(await execa.node('unicorns'));
-expectType<ExecaReturnValue<string>>(
-	await execa.node('unicorns', {encoding: 'utf8'})
+expectType<ExecaChildProcess>(execaNode('unicorns'));
+expectType<ExecaReturnValue>(await execaNode('unicorns'));
+expectType<ExecaReturnValue>(
+	await execaNode('unicorns', {encoding: 'utf8'}),
 );
-expectType<ExecaReturnValue<Buffer>>(await execa.node('unicorns', {encoding: null}));
-expectType<ExecaReturnValue<string>>(
-	await execa.node('unicorns', ['foo'], {encoding: 'utf8'})
+expectType<ExecaReturnValue<Buffer>>(await execaNode('unicorns', {encoding: null}));
+expectType<ExecaReturnValue>(
+	await execaNode('unicorns', ['foo'], {encoding: 'utf8'}),
 );
 expectType<ExecaReturnValue<Buffer>>(
-	await execa.node('unicorns', ['foo'], {encoding: null})
+	await execaNode('unicorns', ['foo'], {encoding: null}),
 );
