@@ -1,15 +1,17 @@
-import path from 'path';
+import path from 'node:path';
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
 import test from 'ava';
-import execa from '..';
+import {execa, execaSync, execaCommand, execaCommandSync} from '../index.js';
 
-process.env.PATH = path.join(__dirname, 'fixtures') + path.delimiter + process.env.PATH;
+process.env.PATH = fileURLToPath(new URL('fixtures', import.meta.url)) + path.delimiter + process.env.PATH;
 
 const command = async (t, expected, ...args) => {
-	const {command: failCommand} = await t.throwsAsync(execa('fail', args));
-	t.is(failCommand, `fail${expected}`);
+	const {command: failCommand} = await t.throwsAsync(execa('fail.js', args));
+	t.is(failCommand, `fail.js${expected}`);
 
-	const {command} = await execa('noop', args);
-	t.is(command, `noop${expected}`);
+	const {command} = await execa('noop.js', args);
+	t.is(command, `noop.js${expected}`);
 };
 
 command.title = (message, expected) => `command is: ${JSON.stringify(expected)}`;
@@ -18,47 +20,70 @@ test(command, ' foo bar', 'foo', 'bar');
 test(command, ' baz quz', 'baz', 'quz');
 test(command, '');
 
+const testEscapedCommand = async (t, expected, args) => {
+	const {escapedCommand: failEscapedCommand} = await t.throwsAsync(execa('fail.js', args));
+	t.is(failEscapedCommand, `fail.js ${expected}`);
+
+	const {escapedCommand: failEscapedCommandSync} = t.throws(() => {
+		execaSync('fail.js', args);
+	});
+	t.is(failEscapedCommandSync, `fail.js ${expected}`);
+
+	const {escapedCommand} = await execa('noop.js', args);
+	t.is(escapedCommand, `noop.js ${expected}`);
+
+	const {escapedCommand: escapedCommandSync} = execaSync('noop.js', args);
+	t.is(escapedCommandSync, `noop.js ${expected}`);
+};
+
+testEscapedCommand.title = (message, expected) => `escapedCommand is: ${JSON.stringify(expected)}`;
+
+test(testEscapedCommand, 'foo bar', ['foo', 'bar']);
+test(testEscapedCommand, '"foo bar"', ['foo bar']);
+test(testEscapedCommand, '"\\"foo\\""', ['"foo"']);
+test(testEscapedCommand, '"*"', ['*']);
+
 test('allow commands with spaces and no array arguments', async t => {
-	const {stdout} = await execa('command with space');
+	const {stdout} = await execa('command with space.js');
 	t.is(stdout, '');
 });
 
 test('allow commands with spaces and array arguments', async t => {
-	const {stdout} = await execa('command with space', ['foo', 'bar']);
+	const {stdout} = await execa('command with space.js', ['foo', 'bar']);
 	t.is(stdout, 'foo\nbar');
 });
 
-test('execa.command()', async t => {
-	const {stdout} = await execa.command('node test/fixtures/echo foo bar');
+test('execaCommand()', async t => {
+	const {stdout} = await execaCommand('node test/fixtures/echo.js foo bar');
 	t.is(stdout, 'foo\nbar');
 });
 
-test('execa.command() ignores consecutive spaces', async t => {
-	const {stdout} = await execa.command('node test/fixtures/echo foo    bar');
+test('execaCommand() ignores consecutive spaces', async t => {
+	const {stdout} = await execaCommand('node test/fixtures/echo.js foo    bar');
 	t.is(stdout, 'foo\nbar');
 });
 
-test('execa.command() allows escaping spaces in commands', async t => {
-	const {stdout} = await execa.command('command\\ with\\ space foo bar');
+test('execaCommand() allows escaping spaces in commands', async t => {
+	const {stdout} = await execaCommand('command\\ with\\ space.js foo bar');
 	t.is(stdout, 'foo\nbar');
 });
 
-test('execa.command() allows escaping spaces in arguments', async t => {
-	const {stdout} = await execa.command('node test/fixtures/echo foo\\ bar');
+test('execaCommand() allows escaping spaces in arguments', async t => {
+	const {stdout} = await execaCommand('node test/fixtures/echo.js foo\\ bar');
 	t.is(stdout, 'foo bar');
 });
 
-test('execa.command() escapes other whitespaces', async t => {
-	const {stdout} = await execa.command('node test/fixtures/echo foo\tbar');
+test('execaCommand() escapes other whitespaces', async t => {
+	const {stdout} = await execaCommand('node test/fixtures/echo.js foo\tbar');
 	t.is(stdout, 'foo\tbar');
 });
 
-test('execa.command() trims', async t => {
-	const {stdout} = await execa.command('  node test/fixtures/echo foo bar  ');
+test('execaCommand() trims', async t => {
+	const {stdout} = await execaCommand('  node test/fixtures/echo.js foo bar  ');
 	t.is(stdout, 'foo\nbar');
 });
 
-test('execa.command.sync()', t => {
-	const {stdout} = execa.commandSync('node test/fixtures/echo foo bar');
+test('execaCommandSync()', t => {
+	const {stdout} = execaCommandSync('node test/fixtures/echo.js foo bar');
 	t.is(stdout, 'foo\nbar');
 });
