@@ -1,24 +1,29 @@
-import path from 'path';
-import childProcess from 'child_process';
+import path from 'node:path';
+import process from 'node:process';
+import childProcess from 'node:child_process';
+import {fileURLToPath} from 'node:url';
+import {promisify} from 'node:util';
 import test from 'ava';
-import execa from '..';
+import {execa, execaSync} from '../index.js';
 
-process.env.PATH = path.join(__dirname, 'fixtures') + path.delimiter + process.env.PATH;
+const pExec = promisify(childProcess.exec);
+
+process.env.PATH = fileURLToPath(new URL('fixtures', import.meta.url)) + path.delimiter + process.env.PATH;
 
 const TIMEOUT_REGEXP = /timed out after/;
 
 const getExitRegExp = exitMessage => new RegExp(`failed with exit code ${exitMessage}`);
 
 test('stdout/stderr/all available on errors', async t => {
-	const {stdout, stderr, all} = await t.throwsAsync(execa('exit', ['2'], {all: true}), {message: getExitRegExp('2')});
+	const {stdout, stderr, all} = await t.throwsAsync(execa('exit.js', ['2'], {all: true}), {message: getExitRegExp('2')});
 	t.is(typeof stdout, 'string');
 	t.is(typeof stderr, 'string');
 	t.is(typeof all, 'string');
 });
 
-const WRONG_COMMAND = process.platform === 'win32' ?
-	'\'wrong\' is not recognized as an internal or external command,\r\noperable program or batch file.' :
-	'';
+const WRONG_COMMAND = process.platform === 'win32'
+	? '\'wrong\' is not recognized as an internal or external command,\r\noperable program or batch file.'
+	: '';
 
 test('stdout/stderr/all on process errors', async t => {
 	const {stdout, stderr, all} = await t.throwsAsync(execa('wrong command', {all: true}));
@@ -29,7 +34,7 @@ test('stdout/stderr/all on process errors', async t => {
 
 test('stdout/stderr/all on process errors, in sync mode', t => {
 	const {stdout, stderr, all} = t.throws(() => {
-		execa.sync('wrong command');
+		execaSync('wrong command');
 	});
 	t.is(stdout, '');
 	t.is(stderr, WRONG_COMMAND);
@@ -37,13 +42,13 @@ test('stdout/stderr/all on process errors, in sync mode', t => {
 });
 
 test('exitCode is 0 on success', async t => {
-	const {exitCode} = await execa('noop', ['foo']);
+	const {exitCode} = await execa('noop.js', ['foo']);
 	t.is(exitCode, 0);
 });
 
-const testExitCode = async (t, num) => {
-	const {exitCode} = await t.throwsAsync(execa('exit', [`${num}`]), {message: getExitRegExp(num)});
-	t.is(exitCode, num);
+const testExitCode = async (t, number) => {
+	const {exitCode} = await t.throwsAsync(execa('exit.js', [`${number}`]), {message: getExitRegExp(number)});
+	t.is(exitCode, number);
 };
 
 test('exitCode is 2', testExitCode, 2);
@@ -51,44 +56,47 @@ test('exitCode is 3', testExitCode, 3);
 test('exitCode is 4', testExitCode, 4);
 
 test('error.message contains the command', async t => {
-	await t.throwsAsync(execa('exit', ['2', 'foo', 'bar']), {message: /exit 2 foo bar/});
+	await t.throwsAsync(execa('exit.js', ['2', 'foo', 'bar']), {message: /exit.js 2 foo bar/});
 });
 
 test('error.message contains stdout/stderr if available', async t => {
-	const {message} = await t.throwsAsync(execa('echo-fail'));
+	const {message} = await t.throwsAsync(execa('echo-fail.js'));
 	t.true(message.includes('stderr'));
 	t.true(message.includes('stdout'));
 });
 
 test('error.message does not contain stdout/stderr if not available', async t => {
-	const {message} = await t.throwsAsync(execa('echo-fail', {stdio: 'ignore'}));
+	const {message} = await t.throwsAsync(execa('echo-fail.js', {stdio: 'ignore'}));
 	t.false(message.includes('stderr'));
 	t.false(message.includes('stdout'));
 });
 
 test('error.shortMessage does not contain stdout/stderr', async t => {
-	const {shortMessage} = await t.throwsAsync(execa('echo-fail'));
+	const {shortMessage} = await t.throwsAsync(execa('echo-fail.js'));
 	t.false(shortMessage.includes('stderr'));
 	t.false(shortMessage.includes('stdout'));
 });
 
 test('Original error.message is kept', async t => {
-	const {originalMessage} = await t.throwsAsync(execa('wrong command'));
-	t.is(originalMessage, 'spawn wrong command ENOENT');
+	const {originalMessage} = await t.throwsAsync(execa('noop.js', {cwd: 1}));
+	// On Node >=14.18.0, the error message is
+	// `The "options.cwd" property must be of type string or an instance of Buffer or URL. Received type number (1)`
+	t.true(originalMessage.startsWith('The "options.cwd" property must be of type string'));
+	t.true(originalMessage.includes('. Received type number'));
 });
 
 test('failed is false on success', async t => {
-	const {failed} = await execa('noop', ['foo']);
+	const {failed} = await execa('noop.js', ['foo']);
 	t.false(failed);
 });
 
 test('failed is true on failure', async t => {
-	const {failed} = await t.throwsAsync(execa('exit', ['2']));
+	const {failed} = await t.throwsAsync(execa('exit.js', ['2']));
 	t.true(failed);
 });
 
 test('error.killed is true if process was killed directly', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 
 	subprocess.kill();
 
@@ -97,7 +105,7 @@ test('error.killed is true if process was killed directly', async t => {
 });
 
 test('error.killed is false if process was killed indirectly', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 
 	process.kill(subprocess.pid, 'SIGINT');
 
@@ -108,12 +116,12 @@ test('error.killed is false if process was killed indirectly', async t => {
 });
 
 test('result.killed is false if not killed', async t => {
-	const {killed} = await execa('noop');
+	const {killed} = await execa('noop.js');
 	t.false(killed);
 });
 
 test('result.killed is false if not killed, in sync mode', t => {
-	const {killed} = execa.sync('noop');
+	const {killed} = execaSync('noop.js');
 	t.false(killed);
 });
 
@@ -124,26 +132,26 @@ test('result.killed is false on process error', async t => {
 
 test('result.killed is false on process error, in sync mode', t => {
 	const {killed} = t.throws(() => {
-		execa.sync('wrong command');
+		execaSync('wrong command');
 	});
 	t.false(killed);
 });
 
 if (process.platform === 'darwin') {
-	test.cb('sanity check: child_process.exec also has killed.false if killed indirectly', t => {
-		const {pid} = childProcess.exec('noop', error => {
-			t.truthy(error);
-			t.false(error.killed);
-			t.end();
-		});
+	test('sanity check: child_process.exec also has killed.false if killed indirectly', async t => {
+		const promise = pExec('noop.js');
 
-		process.kill(pid, 'SIGINT');
+		process.kill(promise.child.pid, 'SIGINT');
+
+		const error = await t.throwsAsync(promise);
+		t.truthy(error);
+		t.false(error.killed);
 	});
 }
 
 if (process.platform !== 'win32') {
 	test('error.signal is SIGINT', async t => {
-		const subprocess = execa('noop');
+		const subprocess = execa('noop.js');
 
 		process.kill(subprocess.pid, 'SIGINT');
 
@@ -152,7 +160,7 @@ if (process.platform !== 'win32') {
 	});
 
 	test('error.signalDescription is defined', async t => {
-		const subprocess = execa('noop');
+		const subprocess = execa('noop.js');
 
 		process.kill(subprocess.pid, 'SIGINT');
 
@@ -161,7 +169,7 @@ if (process.platform !== 'win32') {
 	});
 
 	test('error.signal is SIGTERM', async t => {
-		const subprocess = execa('noop');
+		const subprocess = execa('noop.js');
 
 		process.kill(subprocess.pid, 'SIGTERM');
 
@@ -170,12 +178,12 @@ if (process.platform !== 'win32') {
 	});
 
 	test('custom error.signal', async t => {
-		const {signal} = await t.throwsAsync(execa('noop', {killSignal: 'SIGHUP', timeout: 1, message: TIMEOUT_REGEXP}));
+		const {signal} = await t.throwsAsync(execa('noop.js', {killSignal: 'SIGHUP', timeout: 1, message: TIMEOUT_REGEXP}));
 		t.is(signal, 'SIGHUP');
 	});
 
 	test('exitCode is undefined on signal termination', async t => {
-		const subprocess = execa('noop');
+		const subprocess = execa('noop.js');
 
 		process.kill(subprocess.pid);
 
@@ -185,26 +193,26 @@ if (process.platform !== 'win32') {
 }
 
 test('result.signal is undefined for successful execution', async t => {
-	const {signal} = await execa('noop');
+	const {signal} = await execa('noop.js');
 	t.is(signal, undefined);
 });
 
 test('result.signal is undefined if process failed, but was not killed', async t => {
-	const {signal} = await t.throwsAsync(execa('exit', [2]), {message: getExitRegExp('2')});
+	const {signal} = await t.throwsAsync(execa('exit.js', [2]), {message: getExitRegExp('2')});
 	t.is(signal, undefined);
 });
 
 test('result.signalDescription is undefined for successful execution', async t => {
-	const {signalDescription} = await execa('noop');
+	const {signalDescription} = await execa('noop.js');
 	t.is(signalDescription, undefined);
 });
 
 test('error.code is undefined on success', async t => {
-	const {code} = await execa('noop');
+	const {code} = await execa('noop.js');
 	t.is(code, undefined);
 });
 
 test('error.code is defined on failure if applicable', async t => {
-	const {code} = await t.throwsAsync(execa('invalid'));
-	t.is(code, 'ENOENT');
+	const {code} = await t.throwsAsync(execa('noop.js', {cwd: 1}));
+	t.is(code, 'ERR_INVALID_ARG_TYPE');
 });
