@@ -1,15 +1,17 @@
-import path from 'path';
+import path from 'node:path';
+import process from 'node:process';
+import {fileURLToPath} from 'node:url';
 import test from 'ava';
-import pEvent from 'p-event';
+import {pEvent} from 'p-event';
 import isRunning from 'is-running';
-import execa from '..';
+import {execa, execaSync} from '../index.js';
 
-process.env.PATH = path.join(__dirname, 'fixtures') + path.delimiter + process.env.PATH;
+process.env.PATH = fileURLToPath(new URL('fixtures', import.meta.url)) + path.delimiter + process.env.PATH;
 
 const TIMEOUT_REGEXP = /timed out after/;
 
 test('kill("SIGKILL") should terminate cleanly', async t => {
-	const subprocess = execa('node', ['./test/fixtures/no-killable'], {stdio: ['ipc']});
+	const subprocess = execa('node', ['./test/fixtures/no-killable.js'], {stdio: ['ipc']});
 	await pEvent(subprocess, 'message');
 
 	subprocess.kill('SIGKILL');
@@ -22,7 +24,7 @@ test('kill("SIGKILL") should terminate cleanly', async t => {
 // Therefore, this feature and those tests do not make sense on Windows.
 if (process.platform !== 'win32') {
 	test('`forceKillAfterTimeout: false` should not kill after a timeout', async t => {
-		const subprocess = execa('node', ['./test/fixtures/no-killable'], {stdio: ['ipc']});
+		const subprocess = execa('node', ['./test/fixtures/no-killable.js'], {stdio: ['ipc']});
 		await pEvent(subprocess, 'message');
 
 		subprocess.kill('SIGTERM', {forceKillAfterTimeout: false});
@@ -32,7 +34,7 @@ if (process.platform !== 'win32') {
 	});
 
 	test('`forceKillAfterTimeout: number` should kill after a timeout', async t => {
-		const subprocess = execa('node', ['./test/fixtures/no-killable'], {stdio: ['ipc']});
+		const subprocess = execa('node', ['./test/fixtures/no-killable.js'], {stdio: ['ipc']});
 		await pEvent(subprocess, 'message');
 
 		subprocess.kill('SIGTERM', {forceKillAfterTimeout: 50});
@@ -42,7 +44,7 @@ if (process.platform !== 'win32') {
 	});
 
 	test('`forceKillAfterTimeout: true` should kill after a timeout', async t => {
-		const subprocess = execa('node', ['./test/fixtures/no-killable'], {stdio: ['ipc']});
+		const subprocess = execa('node', ['./test/fixtures/no-killable.js'], {stdio: ['ipc']});
 		await pEvent(subprocess, 'message');
 
 		subprocess.kill('SIGTERM', {forceKillAfterTimeout: true});
@@ -52,7 +54,7 @@ if (process.platform !== 'win32') {
 	});
 
 	test('kill() with no arguments should kill after a timeout', async t => {
-		const subprocess = execa('node', ['./test/fixtures/no-killable'], {stdio: ['ipc']});
+		const subprocess = execa('node', ['./test/fixtures/no-killable.js'], {stdio: ['ipc']});
 		await pEvent(subprocess, 'message');
 
 		subprocess.kill();
@@ -63,38 +65,38 @@ if (process.platform !== 'win32') {
 
 	test('`forceKillAfterTimeout` should not be NaN', t => {
 		t.throws(() => {
-			execa('noop').kill('SIGTERM', {forceKillAfterTimeout: NaN});
+			execa('noop.js').kill('SIGTERM', {forceKillAfterTimeout: Number.NaN});
 		}, {instanceOf: TypeError, message: /non-negative integer/});
 	});
 
 	test('`forceKillAfterTimeout` should not be negative', t => {
 		t.throws(() => {
-			execa('noop').kill('SIGTERM', {forceKillAfterTimeout: -1});
+			execa('noop.js').kill('SIGTERM', {forceKillAfterTimeout: -1});
 		}, {instanceOf: TypeError, message: /non-negative integer/});
 	});
 }
 
 test('execa() returns a promise with kill()', t => {
-	const {kill} = execa('noop', ['foo']);
+	const {kill} = execa('noop.js', ['foo']);
 	t.is(typeof kill, 'function');
 });
 
 test('timeout kills the process if it times out', async t => {
-	const {killed, timedOut} = await t.throwsAsync(execa('noop', {timeout: 1}), TIMEOUT_REGEXP);
+	const {killed, timedOut} = await t.throwsAsync(execa('noop.js', {timeout: 1}), {message: TIMEOUT_REGEXP});
 	t.false(killed);
 	t.true(timedOut);
 });
 
 test('timeout kills the process if it times out, in sync mode', async t => {
 	const {killed, timedOut} = await t.throws(() => {
-		execa.sync('noop', {timeout: 1, message: TIMEOUT_REGEXP});
+		execaSync('noop.js', {timeout: 1, message: TIMEOUT_REGEXP});
 	});
 	t.false(killed);
 	t.true(timedOut);
 });
 
 test('timeout does not kill the process if it does not time out', async t => {
-	const {timedOut} = await execa('delay', ['500'], {timeout: 1e8});
+	const {timedOut} = await execa('delay.js', ['500'], {timeout: 1e8});
 	t.false(timedOut);
 });
 
@@ -102,34 +104,34 @@ const INVALID_TIMEOUT_REGEXP = /`timeout` option to be a non-negative integer/;
 
 test('timeout must not be negative', async t => {
 	await t.throws(() => {
-		execa('noop', {timeout: -1});
-	}, INVALID_TIMEOUT_REGEXP);
+		execa('noop.js', {timeout: -1});
+	}, {message: INVALID_TIMEOUT_REGEXP});
 });
 
 test('timeout must be an integer', async t => {
 	await t.throws(() => {
-		execa('noop', {timeout: false});
-	}, INVALID_TIMEOUT_REGEXP);
+		execa('noop.js', {timeout: false});
+	}, {message: INVALID_TIMEOUT_REGEXP});
 });
 
 test('timedOut is false if timeout is undefined', async t => {
-	const {timedOut} = await execa('noop');
+	const {timedOut} = await execa('noop.js');
 	t.false(timedOut);
 });
 
 test('timedOut is false if timeout is 0', async t => {
-	const {timedOut} = await execa('noop', {timeout: 0});
+	const {timedOut} = await execa('noop.js', {timeout: 0});
 	t.false(timedOut);
 });
 
 test('timedOut is false if timeout is undefined and exit code is 0 in sync mode', t => {
-	const {timedOut} = execa.sync('noop');
+	const {timedOut} = execaSync('noop.js');
 	t.false(timedOut);
 });
 
 // When child process exits before parent process
 const spawnAndExit = async (t, cleanup, detached) => {
-	await t.notThrowsAsync(execa('sub-process-exit', [cleanup, detached]));
+	await t.notThrowsAsync(execa('sub-process-exit.js', [cleanup, detached]));
 };
 
 test('spawnAndExit', spawnAndExit, false, false);
@@ -138,8 +140,8 @@ test('spawnAndExit detached', spawnAndExit, false, true);
 test('spawnAndExit cleanup detached', spawnAndExit, true, true);
 
 // When parent process exits before child process
-const spawnAndKill = async (t, signal, cleanup, detached, isKilled) => {
-	const subprocess = execa('sub-process', [cleanup, detached], {stdio: ['ignore', 'ignore', 'ignore', 'ipc']});
+const spawnAndKill = async (t, [signal, cleanup, detached, isKilled]) => {
+	const subprocess = execa('sub-process.js', [cleanup, detached], {stdio: ['ignore', 'ignore', 'ignore', 'ipc']});
 
 	const pid = await pEvent(subprocess, 'message');
 	t.true(Number.isInteger(pid));
@@ -164,21 +166,21 @@ const spawnAndKill = async (t, signal, cleanup, detached, isKilled) => {
 // With `options.cleanup`, subprocesses are always killed
 //   - `options.cleanup` with SIGKILL is a noop, since it cannot be handled
 const exitIfWindows = process.platform === 'win32';
-test('spawnAndKill SIGTERM', spawnAndKill, 'SIGTERM', false, false, exitIfWindows);
-test('spawnAndKill SIGKILL', spawnAndKill, 'SIGKILL', false, false, exitIfWindows);
-test('spawnAndKill cleanup SIGTERM', spawnAndKill, 'SIGTERM', true, false, true);
-test('spawnAndKill cleanup SIGKILL', spawnAndKill, 'SIGKILL', true, false, exitIfWindows);
-test('spawnAndKill detached SIGTERM', spawnAndKill, 'SIGTERM', false, true, false);
-test('spawnAndKill detached SIGKILL', spawnAndKill, 'SIGKILL', false, true, false);
-test('spawnAndKill cleanup detached SIGTERM', spawnAndKill, 'SIGTERM', true, true, false);
-test('spawnAndKill cleanup detached SIGKILL', spawnAndKill, 'SIGKILL', true, true, false);
+test('spawnAndKill SIGTERM', spawnAndKill, ['SIGTERM', false, false, exitIfWindows]);
+test('spawnAndKill SIGKILL', spawnAndKill, ['SIGKILL', false, false, exitIfWindows]);
+test('spawnAndKill cleanup SIGTERM', spawnAndKill, ['SIGTERM', true, false, true]);
+test('spawnAndKill cleanup SIGKILL', spawnAndKill, ['SIGKILL', true, false, exitIfWindows]);
+test('spawnAndKill detached SIGTERM', spawnAndKill, ['SIGTERM', false, true, false]);
+test('spawnAndKill detached SIGKILL', spawnAndKill, ['SIGKILL', false, true, false]);
+test('spawnAndKill cleanup detached SIGTERM', spawnAndKill, ['SIGTERM', true, true, false]);
+test('spawnAndKill cleanup detached SIGKILL', spawnAndKill, ['SIGKILL', true, true, false]);
 
 // See #128
 test('removes exit handler on exit', async t => {
-	// FIXME: This relies on `signal-exit` internals
+	// @todo this relies on `signal-exit` internals
 	const emitter = process.__signal_exit_emitter__;
 
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	const listener = emitter.listeners('exit').pop();
 
 	await new Promise((resolve, reject) => {
@@ -197,49 +199,49 @@ test('cancel method kills the subprocess', t => {
 });
 
 test('result.isCanceled is false when spawned.cancel() isn\'t called (success)', async t => {
-	const {isCanceled} = await execa('noop');
+	const {isCanceled} = await execa('noop.js');
 	t.false(isCanceled);
 });
 
 test('result.isCanceled is false when spawned.cancel() isn\'t called (failure)', async t => {
-	const {isCanceled} = await t.throwsAsync(execa('fail'));
+	const {isCanceled} = await t.throwsAsync(execa('fail.js'));
 	t.false(isCanceled);
 });
 
 test('result.isCanceled is false when spawned.cancel() isn\'t called in sync mode (success)', t => {
-	const {isCanceled} = execa.sync('noop');
+	const {isCanceled} = execaSync('noop.js');
 	t.false(isCanceled);
 });
 
 test('result.isCanceled is false when spawned.cancel() isn\'t called in sync mode (failure)', t => {
 	const {isCanceled} = t.throws(() => {
-		execa.sync('fail');
+		execaSync('fail.js');
 	});
 	t.false(isCanceled);
 });
 
 test('calling cancel method throws an error with message "Command was canceled"', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.cancel();
 	await t.throwsAsync(subprocess, {message: /Command was canceled/});
 });
 
 test('error.isCanceled is true when cancel method is used', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.cancel();
 	const {isCanceled} = await t.throwsAsync(subprocess);
 	t.true(isCanceled);
 });
 
 test('error.isCanceled is false when kill method is used', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.kill();
 	const {isCanceled} = await t.throwsAsync(subprocess);
 	t.false(isCanceled);
 });
 
 test('calling cancel method twice should show the same behaviour as calling it once', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.cancel();
 	subprocess.cancel();
 	const {isCanceled} = await t.throwsAsync(subprocess);
@@ -248,15 +250,62 @@ test('calling cancel method twice should show the same behaviour as calling it o
 });
 
 test('calling cancel method on a successfully completed process does not make result.isCanceled true', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	const {isCanceled} = await subprocess;
 	subprocess.cancel();
 	t.false(isCanceled);
 });
 
 test('calling cancel method on a process which has been killed does not make error.isCanceled true', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.kill();
 	const {isCanceled} = await t.throwsAsync(subprocess);
 	t.false(isCanceled);
 });
+
+if (globalThis.AbortController !== undefined) {
+	test('calling abort throws an error with message "Command was canceled"', async t => {
+		const abortController = new AbortController();
+		const subprocess = execa('noop.js', [], {signal: abortController.signal});
+		abortController.abort();
+		await t.throwsAsync(subprocess, {message: /Command was canceled/});
+	});
+
+	test('calling abort twice should show the same behaviour as calling it once', async t => {
+		const abortController = new AbortController();
+		const subprocess = execa('noop.js', [], {signal: abortController.signal});
+		abortController.abort();
+		abortController.abort();
+		const {isCanceled} = await t.throwsAsync(subprocess);
+		t.true(isCanceled);
+		t.true(subprocess.killed);
+	});
+
+	test('calling abort on a successfully completed process does not make result.isCanceled true', async t => {
+		const abortController = new AbortController();
+		const subprocess = execa('noop.js', [], {signal: abortController.signal});
+		const {isCanceled} = await subprocess;
+		abortController.abort();
+		t.false(isCanceled);
+	});
+
+	test('calling cancel after abort should show the same behaviour as only calling cancel', async t => {
+		const abortController = new AbortController();
+		const subprocess = execa('noop.js', [], {signal: abortController.signal});
+		abortController.abort();
+		subprocess.cancel();
+		const {isCanceled} = await t.throwsAsync(subprocess);
+		t.true(isCanceled);
+		t.true(subprocess.killed);
+	});
+
+	test('calling abort after cancel should show the same behaviour as only calling cancel', async t => {
+		const abortController = new AbortController();
+		const subprocess = execa('noop.js', [], {signal: abortController.signal});
+		subprocess.cancel();
+		abortController.abort();
+		const {isCanceled} = await t.throwsAsync(subprocess);
+		t.true(isCanceled);
+		t.true(subprocess.killed);
+	});
+}
