@@ -1,19 +1,19 @@
-import util from 'util';
+import {inspect} from 'node:util';
 import test from 'ava';
-import normalizeStdio from '../lib/stdio';
+import {normalizeStdio, normalizeStdioNode} from '../lib/stdio.js';
 
 const macro = (t, input, expected, func) => {
 	if (expected instanceof Error) {
 		t.throws(() => {
 			normalizeStdio(input);
-		}, expected.message);
+		}, {message: expected.message});
 		return;
 	}
 
 	t.deepEqual(func(input), expected);
 };
 
-const macroTitle = name => (title, input) => `${name} ${(util.inspect(input))}`;
+const macroTitle = name => (title, input) => `${name} ${(inspect(input))}`;
 
 const stdioMacro = (...args) => macro(...args, normalizeStdio);
 stdioMacro.title = macroTitle('execa()');
@@ -47,8 +47,8 @@ test(stdioMacro, {stdin: 'inherit', stdio: ['pipe']}, new Error('It\'s not possi
 test(stdioMacro, {stdin: 'inherit', stdio: [undefined, 'pipe']}, new Error('It\'s not possible to provide `stdio` in combination with one of `stdin`, `stdout`, `stderr`'));
 test(stdioMacro, {stdin: 0, stdio: 'pipe'}, new Error('It\'s not possible to provide `stdio` in combination with one of `stdin`, `stdout`, `stderr`'));
 
-const forkMacro = (...args) => macro(...args, normalizeStdio.node);
-forkMacro.title = macroTitle('execa.fork()');
+const forkMacro = (...args) => macro(...args, normalizeStdioNode);
+forkMacro.title = macroTitle('execaNode()');
 
 test(forkMacro, undefined, [undefined, undefined, undefined, 'ipc']);
 test(forkMacro, {stdio: 'ignore'}, ['ignore', 'ignore', 'ignore', 'ipc']);
