@@ -1,62 +1,65 @@
-import path from 'path';
-import fs from 'fs';
-import Stream from 'stream';
+import {Buffer} from 'node:buffer';
+import path from 'node:path';
+import process from 'node:process';
+import fs from 'node:fs';
+import {fileURLToPath} from 'node:url';
+import Stream from 'node:stream';
 import test from 'ava';
 import getStream from 'get-stream';
 import tempfile from 'tempfile';
-import execa from '..';
+import {execa, execaSync} from '../index.js';
 
-process.env.PATH = path.join(__dirname, 'fixtures') + path.delimiter + process.env.PATH;
+process.env.PATH = fileURLToPath(new URL('fixtures', import.meta.url)) + path.delimiter + process.env.PATH;
 
 test('buffer', async t => {
-	const {stdout} = await execa('noop', ['foo'], {encoding: null});
+	const {stdout} = await execa('noop.js', ['foo'], {encoding: null});
 	t.true(Buffer.isBuffer(stdout));
 	t.is(stdout.toString(), 'foo');
 });
 
 test('pass `stdout` to a file descriptor', async t => {
 	const file = tempfile('.txt');
-	await execa('test/fixtures/noop', ['foo bar'], {stdout: fs.openSync(file, 'w')});
+	await execa('test/fixtures/noop.js', ['foo bar'], {stdout: fs.openSync(file, 'w')});
 	t.is(fs.readFileSync(file, 'utf8'), 'foo bar\n');
 });
 
 test('pass `stderr` to a file descriptor', async t => {
 	const file = tempfile('.txt');
-	await execa('test/fixtures/noop-err', ['foo bar'], {stderr: fs.openSync(file, 'w')});
+	await execa('test/fixtures/noop-err.js', ['foo bar'], {stderr: fs.openSync(file, 'w')});
 	t.is(fs.readFileSync(file, 'utf8'), 'foo bar\n');
 });
 
 test.serial('result.all shows both `stdout` and `stderr` intermixed', async t => {
-	const {all} = await execa('noop-132', {all: true});
+	const {all} = await execa('noop-132.js', {all: true});
 	t.is(all, '132');
 });
 
 test('result.all is undefined unless opts.all is true', async t => {
-	const {all} = await execa('noop');
+	const {all} = await execa('noop.js');
 	t.is(all, undefined);
 });
 
 test('stdout/stderr/all are undefined if ignored', async t => {
-	const {stdout, stderr, all} = await execa('noop', {stdio: 'ignore', all: true});
+	const {stdout, stderr, all} = await execa('noop.js', {stdio: 'ignore', all: true});
 	t.is(stdout, undefined);
 	t.is(stderr, undefined);
 	t.is(all, undefined);
 });
 
 test('stdout/stderr/all are undefined if ignored in sync mode', t => {
-	const {stdout, stderr, all} = execa.sync('noop', {stdio: 'ignore', all: true});
+	const {stdout, stderr, all} = execaSync('noop.js', {stdio: 'ignore', all: true});
 	t.is(stdout, undefined);
 	t.is(stderr, undefined);
 	t.is(all, undefined);
 });
 
 test('input option can be a String', async t => {
-	const {stdout} = await execa('stdin', {input: 'foobar'});
+	const {stdout} = await execa('stdin.js', {input: 'foobar'});
 	t.is(stdout, 'foobar');
 });
 
 test('input option can be a Buffer', async t => {
-	const {stdout} = await execa('stdin', {input: 'testing12'});
+	const {stdout} = await execa('stdin.js', {input: 'testing12'});
 	t.is(stdout, 'testing12');
 });
 
@@ -64,30 +67,31 @@ test('input can be a Stream', async t => {
 	const stream = new Stream.PassThrough();
 	stream.write('howdy');
 	stream.end();
-	const {stdout} = await execa('stdin', {input: stream});
+	const {stdout} = await execa('stdin.js', {input: stream});
 	t.is(stdout, 'howdy');
 });
 
 test('you can write to child.stdin', async t => {
-	const subprocess = execa('stdin');
+	const subprocess = execa('stdin.js');
 	subprocess.stdin.end('unicorns');
-	t.is((await subprocess).stdout, 'unicorns');
+	const {stdout} = await subprocess;
+	t.is(stdout, 'unicorns');
 });
 
 test('input option can be a String - sync', t => {
-	const {stdout} = execa.sync('stdin', {input: 'foobar'});
+	const {stdout} = execaSync('stdin.js', {input: 'foobar'});
 	t.is(stdout, 'foobar');
 });
 
 test('input option can be a Buffer - sync', t => {
-	const {stdout} = execa.sync('stdin', {input: Buffer.from('testing12', 'utf8')});
+	const {stdout} = execaSync('stdin.js', {input: Buffer.from('testing12', 'utf8')});
 	t.is(stdout, 'testing12');
 });
 
 test('opts.stdout:ignore - stdout will not collect data', async t => {
-	const {stdout} = await execa('stdin', {
+	const {stdout} = await execa('stdin.js', {
 		input: 'hello',
-		stdio: [undefined, 'ignore', undefined]
+		stdio: [undefined, 'ignore', undefined],
 	});
 	t.is(stdout, undefined);
 });
@@ -95,31 +99,31 @@ test('opts.stdout:ignore - stdout will not collect data', async t => {
 test('helpful error trying to provide an input stream in sync mode', t => {
 	t.throws(
 		() => {
-			execa.sync('stdin', {input: new Stream.PassThrough()});
+			execaSync('stdin.js', {input: new Stream.PassThrough()});
 		},
-		/The `input` option cannot be a stream in sync mode/
+		{message: /The `input` option cannot be a stream in sync mode/},
 	);
 });
 
 test('maxBuffer affects stdout', async t => {
-	await t.notThrowsAsync(execa('max-buffer', ['stdout', '10'], {maxBuffer: 10}));
-	const {stdout, all} = await t.throwsAsync(execa('max-buffer', ['stdout', '11'], {maxBuffer: 10, all: true}), /max-buffer stdout/);
+	await t.notThrowsAsync(execa('max-buffer.js', ['stdout', '10'], {maxBuffer: 10}));
+	const {stdout, all} = await t.throwsAsync(execa('max-buffer.js', ['stdout', '11'], {maxBuffer: 10, all: true}), {message: /max-buffer.js stdout/});
 	t.is(stdout, '.'.repeat(10));
 	t.is(all, '.'.repeat(10));
 });
 
 test('maxBuffer affects stderr', async t => {
-	await t.notThrowsAsync(execa('max-buffer', ['stderr', '10'], {maxBuffer: 10}));
-	const {stderr, all} = await t.throwsAsync(execa('max-buffer', ['stderr', '11'], {maxBuffer: 10, all: true}), /max-buffer stderr/);
+	await t.notThrowsAsync(execa('max-buffer.js', ['stderr', '10'], {maxBuffer: 10}));
+	const {stderr, all} = await t.throwsAsync(execa('max-buffer.js', ['stderr', '11'], {maxBuffer: 10, all: true}), {message: /max-buffer.js stderr/});
 	t.is(stderr, '.'.repeat(10));
 	t.is(all, '.'.repeat(10));
 });
 
 test('do not buffer stdout when `buffer` set to `false`', async t => {
-	const promise = execa('max-buffer', ['stdout', '10'], {buffer: false});
+	const promise = execa('max-buffer.js', ['stdout', '10'], {buffer: false});
 	const [result, stdout] = await Promise.all([
 		promise,
-		getStream(promise.stdout)
+		getStream(promise.stdout),
 	]);
 
 	t.is(result.stdout, undefined);
@@ -127,10 +131,10 @@ test('do not buffer stdout when `buffer` set to `false`', async t => {
 });
 
 test('do not buffer stderr when `buffer` set to `false`', async t => {
-	const promise = execa('max-buffer', ['stderr', '10'], {buffer: false});
+	const promise = execa('max-buffer.js', ['stderr', '10'], {buffer: false});
 	const [result, stderr] = await Promise.all([
 		promise,
-		getStream(promise.stderr)
+		getStream(promise.stderr),
 	]);
 
 	t.is(result.stderr, undefined);
@@ -138,44 +142,44 @@ test('do not buffer stderr when `buffer` set to `false`', async t => {
 });
 
 test('do not buffer when streaming', async t => {
-	const {stdout} = execa('max-buffer', ['stdout', '21'], {maxBuffer: 10});
+	const {stdout} = execa('max-buffer.js', ['stdout', '21'], {maxBuffer: 10});
 	const result = await getStream(stdout);
 	t.is(result, '....................\n');
 });
 
 test('buffer: false > promise resolves', async t => {
-	await t.notThrowsAsync(execa('noop', {buffer: false}));
+	await t.notThrowsAsync(execa('noop.js', {buffer: false}));
 });
 
 test('buffer: false > promise resolves when output is big but is not pipable', async t => {
-	await t.notThrowsAsync(execa('max-buffer', {buffer: false, stdout: 'ignore'}));
+	await t.notThrowsAsync(execa('max-buffer.js', {buffer: false, stdout: 'ignore'}));
 });
 
 test('buffer: false > promise resolves when output is big and is read', async t => {
-	const subprocess = execa('max-buffer', {buffer: false});
+	const subprocess = execa('max-buffer.js', {buffer: false});
 	subprocess.stdout.resume();
 	subprocess.stderr.resume();
 	await t.notThrowsAsync(subprocess);
 });
 
 test('buffer: false > promise resolves when output is big and "all" is used and is read', async t => {
-	const subprocess = execa('max-buffer', {buffer: false, all: true});
+	const subprocess = execa('max-buffer.js', {buffer: false, all: true});
 	subprocess.all.resume();
 	await t.notThrowsAsync(subprocess);
 });
 
 test('buffer: false > promise rejects when process returns non-zero', async t => {
-	const subprocess = execa('fail', {buffer: false});
+	const subprocess = execa('fail.js', {buffer: false});
 	const {exitCode} = await t.throwsAsync(subprocess);
 	t.is(exitCode, 2);
 });
 
 test('can use all: true with stdout: ignore', async t => {
-	await t.notThrowsAsync(execa('max-buffer', {buffer: false, stdout: 'ignore', all: true}));
+	await t.notThrowsAsync(execa('max-buffer.js', {buffer: false, stdout: 'ignore', all: true}));
 });
 
 test('can use all: true with stderr: ignore', async t => {
-	await t.notThrowsAsync(execa('max-buffer', ['stderr'], {buffer: false, stderr: 'ignore', all: true}));
+	await t.notThrowsAsync(execa('max-buffer.js', ['stderr'], {buffer: false, stderr: 'ignore', all: true}));
 });
 
 const BUFFER_TIMEOUT = 1e3;
@@ -183,12 +187,12 @@ const BUFFER_TIMEOUT = 1e3;
 // On Unix (not Windows), a process won't exit if stdout has not been read.
 if (process.platform !== 'win32') {
 	test.serial('buffer: false > promise does not resolve when output is big and is not read', async t => {
-		const {timedOut} = await t.throwsAsync(execa('max-buffer', {buffer: false, timeout: BUFFER_TIMEOUT}));
+		const {timedOut} = await t.throwsAsync(execa('max-buffer.js', {buffer: false, timeout: BUFFER_TIMEOUT}));
 		t.true(timedOut);
 	});
 
 	test.serial('buffer: false > promise does not resolve when output is big and "all" is used but not read', async t => {
-		const subprocess = execa('max-buffer', {buffer: false, all: true, timeout: BUFFER_TIMEOUT});
+		const subprocess = execa('max-buffer.js', {buffer: false, all: true, timeout: BUFFER_TIMEOUT});
 		subprocess.stdout.resume();
 		subprocess.stderr.resume();
 		const {timedOut} = await t.throwsAsync(subprocess);
