@@ -1,16 +1,19 @@
-import path from 'path';
+import path from 'node:path';
+import process from 'node:process';
+import {fileURLToPath, pathToFileURL} from 'node:url';
 import test from 'ava';
 import isRunning from 'is-running';
 import getNode from 'get-node';
-import execa from '..';
+import semver from 'semver';
+import {execa, execaSync} from '../index.js';
 
-process.env.PATH = path.join(__dirname, 'fixtures') + path.delimiter + process.env.PATH;
+process.env.PATH = fileURLToPath(new URL('fixtures', import.meta.url)) + path.delimiter + process.env.PATH;
 process.env.FOO = 'foo';
 
 const ENOENT_REGEXP = process.platform === 'win32' ? /failed with exit code 1/ : /spawn.* ENOENT/;
 
 test('execa()', async t => {
-	const {stdout} = await execa('noop', ['foo']);
+	const {stdout} = await execa('noop.js', ['foo']);
 	t.is(stdout, 'foo');
 });
 
@@ -26,50 +29,50 @@ if (process.platform === 'win32') {
 	});
 }
 
-test('execa.sync()', t => {
-	const {stdout} = execa.sync('noop', ['foo']);
+test('execaSync()', t => {
+	const {stdout} = execaSync('noop.js', ['foo']);
 	t.is(stdout, 'foo');
 });
 
-test('execa.sync() throws error if written to stderr', t => {
+test('execaSync() throws error if written to stderr', t => {
 	t.throws(() => {
-		execa.sync('foo');
-	}, ENOENT_REGEXP);
+		execaSync('foo');
+	}, {message: ENOENT_REGEXP});
 });
 
 test('skip throwing when using reject option', async t => {
-	const {exitCode} = await execa('fail', {reject: false});
+	const {exitCode} = await execa('fail.js', {reject: false});
 	t.is(exitCode, 2);
 });
 
 test('skip throwing when using reject option in sync mode', t => {
-	const {exitCode} = execa.sync('fail', {reject: false});
+	const {exitCode} = execaSync('fail.js', {reject: false});
 	t.is(exitCode, 2);
 });
 
 test('stripFinalNewline: true', async t => {
-	const {stdout} = await execa('noop', ['foo']);
+	const {stdout} = await execa('noop.js', ['foo']);
 	t.is(stdout, 'foo');
 });
 
 test('stripFinalNewline: false', async t => {
-	const {stdout} = await execa('noop', ['foo'], {stripFinalNewline: false});
+	const {stdout} = await execa('noop.js', ['foo'], {stripFinalNewline: false});
 	t.is(stdout, 'foo\n');
 });
 
 test('stripFinalNewline on failure', async t => {
-	const {stderr} = await t.throwsAsync(execa('noop-throw', ['foo'], {stripFinalNewline: true}));
+	const {stderr} = await t.throwsAsync(execa('noop-throw.js', ['foo'], {stripFinalNewline: true}));
 	t.is(stderr, 'foo');
 });
 
 test('stripFinalNewline in sync mode', t => {
-	const {stdout} = execa.sync('noop', ['foo'], {stripFinalNewline: true});
+	const {stdout} = execaSync('noop.js', ['foo'], {stripFinalNewline: true});
 	t.is(stdout, 'foo');
 });
 
 test('stripFinalNewline in sync mode on failure', t => {
 	const {stderr} = t.throws(() => {
-		execa.sync('noop-throw', ['foo'], {stripFinalNewline: true});
+		execaSync('noop-throw.js', ['foo'], {stripFinalNewline: true});
 	});
 	t.is(stderr, 'foo');
 });
@@ -79,11 +82,11 @@ test('preferLocal: true', async t => {
 });
 
 test('preferLocal: false', async t => {
-	await t.throwsAsync(execa('ava', ['--version'], {preferLocal: false, env: {Path: '', PATH: ''}}), ENOENT_REGEXP);
+	await t.throwsAsync(execa('ava', ['--version'], {preferLocal: false, env: {Path: '', PATH: ''}}), {message: ENOENT_REGEXP});
 });
 
 test('preferLocal: undefined', async t => {
-	await t.throwsAsync(execa('ava', ['--version'], {env: {Path: '', PATH: ''}}), ENOENT_REGEXP);
+	await t.throwsAsync(execa('ava', ['--version'], {env: {Path: '', PATH: ''}}), {message: ENOENT_REGEXP});
 });
 
 test('localDir option', async t => {
@@ -100,30 +103,30 @@ test('execPath option', async t => {
 });
 
 test('stdin errors are handled', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.stdin.emit('error', new Error('test'));
-	await t.throwsAsync(subprocess, /test/);
+	await t.throwsAsync(subprocess, {message: /test/});
 });
 
 test('child process errors are handled', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.emit('error', new Error('test'));
-	await t.throwsAsync(subprocess, /test/);
+	await t.throwsAsync(subprocess, {message: /test/});
 });
 
 test('child process errors rejects promise right away', async t => {
-	const subprocess = execa('noop');
+	const subprocess = execa('noop.js');
 	subprocess.emit('error', new Error('test'));
-	await t.throwsAsync(subprocess, /test/);
+	await t.throwsAsync(subprocess, {message: /test/});
 });
 
 test('execa() returns a promise with pid', t => {
-	const {pid} = execa('noop', ['foo']);
+	const {pid} = execa('noop.js', ['foo']);
 	t.is(typeof pid, 'number');
 });
 
 test('child_process.spawn() propagated errors have correct shape', t => {
-	const subprocess = execa('noop', {uid: -1});
+	const subprocess = execa('noop.js', {uid: -1});
 	t.notThrows(() => {
 		subprocess.catch(() => {});
 		subprocess.unref();
@@ -132,39 +135,41 @@ test('child_process.spawn() propagated errors have correct shape', t => {
 });
 
 test('child_process.spawn() errors are propagated', async t => {
-	const {failed} = await t.throwsAsync(execa('noop', {uid: -1}));
+	const {failed} = await t.throwsAsync(execa('noop.js', {uid: -1}));
 	t.true(failed);
 });
 
 test('child_process.spawnSync() errors are propagated with a correct shape', t => {
 	const {failed} = t.throws(() => {
-		execa.sync('noop', {timeout: -1});
+		execaSync('noop.js', {timeout: -1});
 	});
 	t.true(failed);
 });
 
 test('do not try to consume streams twice', async t => {
-	const subprocess = execa('noop', ['foo']);
-	t.is((await subprocess).stdout, 'foo');
-	t.is((await subprocess).stdout, 'foo');
+	const subprocess = execa('noop.js', ['foo']);
+	const {stdout} = await subprocess;
+	const {stdout: stdout2} = await subprocess;
+	t.is(stdout, 'foo');
+	t.is(stdout2, 'foo');
 });
 
 test('use relative path with \'..\' chars', async t => {
-	const pathViaParentDir = path.join('..', path.basename(path.dirname(__dirname)), 'test', 'fixtures', 'noop');
+	const pathViaParentDir = path.join('..', path.basename(fileURLToPath(new URL('..', import.meta.url))), 'test', 'fixtures', 'noop.js');
 	const {stdout} = await execa(pathViaParentDir, ['foo']);
 	t.is(stdout, 'foo');
 });
 
 if (process.platform !== 'win32') {
 	test('execa() rejects if running non-executable', async t => {
-		const subprocess = execa('non-executable');
+		const subprocess = execa('non-executable.js');
 		await t.throwsAsync(subprocess);
 	});
 
 	test('execa() rejects with correct error and doesn\'t throw if running non-executable with input', async t => {
 		// On Node <12.6.0, `EACCESS` is emitted on `childProcess`.
 		// On Node >=12.6.0, `EPIPE` is emitted on `childProcess.stdin`.
-		await t.throwsAsync(execa('non-executable', {input: 'Hey!'}), /EACCES|EPIPE/);
+		await t.throwsAsync(execa('non-executable.js', {input: 'Hey!'}), {message: /EACCES|EPIPE/});
 	});
 }
 
@@ -182,28 +187,50 @@ if (process.platform !== 'win32') {
 }
 
 test('use environment variables by default', async t => {
-	const {stdout} = await execa('environment');
+	const {stdout} = await execa('environment.js');
 	t.deepEqual(stdout.split('\n'), ['foo', 'undefined']);
 });
 
 test('extend environment variables by default', async t => {
-	const {stdout} = await execa('environment', [], {env: {BAR: 'bar'}});
+	const {stdout} = await execa('environment.js', [], {env: {BAR: 'bar'}});
 	t.deepEqual(stdout.split('\n'), ['foo', 'bar']);
 });
 
 test('do not extend environment with `extendEnv: false`', async t => {
-	const {stdout} = await execa('environment', [], {env: {BAR: 'bar', PATH: process.env.PATH}, extendEnv: false});
+	const {stdout} = await execa('environment.js', [], {env: {BAR: 'bar', PATH: process.env.PATH}, extendEnv: false});
 	t.deepEqual(stdout.split('\n'), ['undefined', 'bar']);
 });
 
+test('can use `options.cwd` as a string', async t => {
+	const cwd = '/';
+	const {stdout} = await execa('node', ['-p', 'process.cwd()'], {cwd});
+	t.is(path.toNamespacedPath(stdout), path.toNamespacedPath(cwd));
+});
+
+if (semver.satisfies(process.version, '^14.18.0 || >=16.4.0')) {
+	test('localDir option can be a URL', async t => {
+		const command = process.platform === 'win32' ? 'echo %PATH%' : 'echo $PATH';
+		const {stdout} = await execa(command, {shell: true, preferLocal: true, localDir: pathToFileURL('/test')});
+		const envPaths = stdout.split(path.delimiter);
+		t.true(envPaths.some(envPath => envPath.endsWith('.bin')));
+	});
+
+	test('can use `options.cwd` as a URL', async t => {
+		const cwd = '/';
+		const cwdUrl = pathToFileURL(cwd);
+		const {stdout} = await execa('node', ['-p', 'process.cwd()'], {cwd: cwdUrl});
+		t.is(path.toNamespacedPath(stdout), path.toNamespacedPath(cwd));
+	});
+}
+
 test('can use `options.shell: true`', async t => {
-	const {stdout} = await execa('node test/fixtures/noop foo', {shell: true});
+	const {stdout} = await execa('node test/fixtures/noop.js foo', {shell: true});
 	t.is(stdout, 'foo');
 });
 
 test('can use `options.shell: string`', async t => {
 	const shell = process.platform === 'win32' ? 'cmd.exe' : '/bin/bash';
-	const {stdout} = await execa('node test/fixtures/noop foo', {shell});
+	const {stdout} = await execa('node test/fixtures/noop.js foo', {shell});
 	t.is(stdout, 'foo');
 });
 
@@ -216,7 +243,7 @@ test('use extend environment with `extendEnv: true` and `shell: true`', async t 
 });
 
 test('detach child process', async t => {
-	const {stdout} = await execa('detach');
+	const {stdout} = await execa('detach.js');
 	const pid = Number(stdout);
 	t.true(Number.isInteger(pid));
 	t.true(isRunning(pid));
